refactor(views): rename route views to View suffix
